requirements.txt
ascii_magic==1.6
keras==2.10.0
Keras_Preprocessing==1.1.2
numpy==1.21.5
Pillow==9.3.0
pyfiglet==0.8.post1
